Fix compiler warnings on 32-bit Windows
- virtual.c: Use int instead of size_t for loop counter
- x86.c: Add explicit float cast in x86_setf
- regress: Suppress unused variable warnings for df/dd

+++ x86.c
@@ -2045,7 +2045,7 @@ x86_setf(dill_stream s, int type, int junk, int dest, double imm)
 
     if (smi->generate_SSE) {
 	if (type == DILL_F) {
-	    a.f = imm;
+	    a.f = (float)imm;
 	    x86_seti(s, EAX, a.i);
 	    BYTE_OUT4(s, 0x66, 0x0f, 0x6e, ModRM(0x3, dest, EAX));
 	} else {
